+++ App/Foundations/CoreModel.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace app\Foundations;
-
-abstract class CoreModel
-{
-    // do something
-}