+++ App/Foundations/common.php
@@ -1,3 +1,7 @@
 <?php
 
 namespace app\Foundations;
+
+class Common
+{
+}